detect_deepfake_app/frontend/requirements.txt
streamlit
requests
matplotlib
pillow